+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecoskills/util/LeaderboardEntry.kt
@@ -4,5 +4,5 @@ import org.bukkit.OfflinePlayer
 
 data class LeaderboardEntry(
     val player: OfflinePlayer,
-    val level: Int
+    val level: Double
 )
